backend/requirements.txt
Django==1.10.6
djangorestframework==3.5.4
django-cors-headers==2.0.2
uWSGI==2.0.14

